% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkforest.R
\name{uninstall_theme}
\alias{uninstall_theme}
\title{Uninstall the darkforest theme}
\usage{
uninstall_theme()
}
\description{
This function removes the darkforest theme from RStudio and deactivates the custom menu theme.
}
